requirements.txt
streamlit
yt-dlp
google-generativeai
faster-whisper
moviepy
opencv-python
mediapipe
numpy
pillow
